refactor: move page back to layout
