+++ apps/af-mvp/src/pages/auth/sign-in/components/sign-in-form.tsx
@@ -1,15 +1,6 @@
 import { useState } from 'react';
 import { SubmitHandler, useForm } from 'react-hook-form';
-import {
-  confirmSignIn,
-  fetchUserAttributes,
-  forgetDevice,
-  rememberDevice,
-  signIn,
-  signOut,
-  signUp,
-} from 'aws-amplify/auth';
-import { randomBytes } from 'crypto';
+import { confirmSignIn, signIn, signUp } from '@mvp/lib/frontend/aws-cognito';
 import { Button, Text } from 'suomifi-ui-components';
 import FormInput from '@shared/components/form/form-input';
 import CustomHeading from '@shared/components/ui/custom-heading';
@@ -109,10 +100,6 @@ function CodeForm({ handleFormSubmit }: FormProps) {
 
 const sleep = () => new Promise(resolve => setTimeout(resolve, 1500));
 
-function getRandomString(bytes: number) {
-  return randomBytes(bytes).toString('base64');
-}
-
 export default function SignIn() {
   const [email, setEmail] = useState('');
   const [isCodeSent, setCodeSent] = useState(false);
@@ -120,61 +107,22 @@ export default function SignIn() {
   const handleEmailSubmit = async (email: string) => {
     setEmail(email);
 
-    const signInAct = async () => {
-      const { isSignedIn, nextStep } = await signIn({
-        username: email,
-        options: {
-          authFlowType: 'CUSTOM_WITHOUT_SRP',
-        },
-      });
-
-      console.log(isSignedIn, nextStep);
-    };
-
-    const signUpAct = async () => {
-      const { isSignUpComplete, userId, nextStep } = await signUp({
-        username: email,
-        password: getRandomString(30), // This is not used but require a value
-        options: {
-          userAttributes: {
-            email,
-          },
-          autoSignIn: { authFlowType: 'CUSTOM_WITHOUT_SRP' },
-        },
-      });
-      console.log(isSignUpComplete, userId, nextStep);
-    };
-
     try {
       // TODO: Check if user exists before signing up
-      await signUpAct();
+      await signUp(email);
     } catch (error) {
       if (!String(error).startsWith('UsernameExistsException')) {
         throw error;
       }
     }
 
-    await signInAct();
+    await signIn(email);
 
     setCodeSent(true);
   };
 
   const handleCodeSubmit = async (code: string) => {
-    // Send the answer to the User Pool
-    // This will throw an error if it’s the 3rd wrong answer
-    try {
-      const { isSignedIn, nextStep } = await confirmSignIn({
-        challengeResponse: code,
-      });
-      console.log('isSignedIn', isSignedIn);
-      console.log('nextStep', nextStep);
-    } catch (error) {
-      console.log('error confirming sign up', error);
-    }
-  };
-
-  const handleLogout = async () => {
-    await signOut();
+    await confirmSignIn(code);
   };
 
   return (
